refactor: move views to separate directory
