+++ app/views/products.py
@@ -1,0 +1,48 @@
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi_filter import FilterDepends
+from fastapi_pagination import Page
+from fastapi_pagination.ext.sqlalchemy import paginate
+from sqlalchemy.orm import Session
+
+from app import crud, schemas
+from app.db import get_db
+
+router = APIRouter(prefix="/products", tags=["Products"])
+
+
+@router.get("", response_model=Page[schemas.ProductFull], tags=["Products"])
+def get_products(
+    filters: schemas.ProductFilter = FilterDepends(schemas.ProductFilter),
+    db: Session = Depends(get_db),
+):
+    return paginate(db, crud.get_products_query(filters=filters))
+
+
+@router.get(
+    "/code/{product_code}",
+    response_model=schemas.ProductFull,
+    tags=["Products"],
+)
+def get_product_by_code(product_code: str, db: Session = Depends(get_db)):
+    db_product = crud.get_product_by_code(db, code=product_code)
+    if not db_product:
+        raise HTTPException(
+            status_code=404,
+            detail=f"Product with code {product_code} not found",
+        )
+    return db_product
+
+
+@router.get(
+    "/{product_id}",
+    response_model=schemas.ProductFull,
+    tags=["Products"],
+)
+def get_product_by_id(product_id: int, db: Session = Depends(get_db)):
+    db_product = crud.get_product_by_id(db, id=product_id)
+    if not db_product:
+        raise HTTPException(
+            status_code=404,
+            detail=f"Product with id {product_id} not found",
+        )
+    return db_product
